refactor(world): split graph.rs by responsibility (TD-026)
Behavior-preserving split of parish-world/src/graph.rs (~1347 LOC) into
focused submodules: schema.rs (types), loader.rs (load/validate),
lookup.rs (name matching/neighbors), pathfinding.rs (BFS/travel time).
Hub mod.rs re-exports the full public API unchanged. All 160 tests pass.
Part of #1200 (TD-026).
